+++ CHANGELOG.md
@@ -1,8 +1,11 @@
 . Minor | - Fix | + Addition | ^ improvement | ! Change | * Refactor | @ Version
 
-# 0.1.1-dev
-`+` Addition: test clinet code 작성 중.
-    WebSocket 통신 및 리소스, 이벤트 구현 중.
+# 0.1.1-dev  
+`+` Addition: test clinet code 작성 중.  
+    WebSocket 통신 및 리소스, 이벤트 구현 중.  
+
+## 2025.08.16  
+`+` Addition: client에서 키보드 입력 이벤트 발생 시 WebSocket으로 연결된 서버로 보낼 수 있도록 기능 구현 중...
     
 # 0.1.0
 `@` Version: first commit